+++ src/js/modal-shop-list.js
@@ -1,5 +1,5 @@
 import {fetchSelectedBook} from "./api"
-import { createMarkupModalAddShopList } from "./createMarkupModalShopList";
+import { createMarkupModalAddShopList, createMarkupModalRemoveShopList } from "./createMarkupModal";
 import { saveInLocal, removeFromLocal } from "./localStorage";
 import "./category";
 import "./bookcard";
@@ -8,12 +8,11 @@ import "./bookcard";
 export const refs = {
     containerModalShopList: document.querySelector('#container-modal-shop-list'),
     bookGallery: document.querySelector('.bookgallery'),
-    // btnAddShopList: document.querySelector('.js-btn-add'),
-    btnRemoveShopList: document.querySelector('.js-btn-remove'),
+    
+    // btnRemoveShopList: document.querySelector('.js-btn-remove'),
     
 };
 
-console.log(refs.bookGallery);
 
 export const API_KEY = 'shoppingList';
 const localStorageBooks = [];
@@ -24,31 +23,42 @@ refs.bookGallery.addEventListener('click', onBookCardClick);
 async function onBookCardClick(event) {
     event.preventDefault();
 
-    if (event.target.classList.contains('js-book-card')) {
+    const targetCard = event.target.closest('.js-book-card');
+
+    if (!targetCard) {
     return;
-    } 
+    }
 
     toggleModal();
 
     refs.containerModalShopList.style.display = 'block';
     document.body.style.overflow = 'hidden';
 
-    const liBook = event.target.closest('.js-book-card');
-    const id = liBook.dataset.id;
-    
+    const id = targetCard.dataset.id;
     const book = await fetchSelectedBook(id);
-
     const markup = createMarkupModalAddShopList(book);
 
     refs.containerModalShopList.innerHTML = markup;
 
-    const btnCloseModalShopList = document.querySelector('.modal-shop-list-close')
+    window.addEventListener('click', onBackdropClick);
+    
+    function onBackdropClick(event) {
+        if (event.target === refs.containerModalShopList) {
+            toggleModal();
+        }
+    }
 
+    const btnCloseModalShopList = document.querySelector('.modal-shop-list-close')
     btnCloseModalShopList.addEventListener('click', toggleModal);
     
     function toggleModal() {
         refs.containerModalShopList.classList.toggle('is-hidden');
+
         window.removeEventListener('keydown', onEscKeyPress);
+        window.removeEventListener('click', onBackdropClick);
+
+        refs.containerModalShopList.style.display = 'none';
+        document.body.style.overflow = 'auto';
     }
     
     window.addEventListener('keydown', onEscKeyPress);
@@ -61,10 +71,14 @@ async function onBookCardClick(event) {
 
     const btnAddShopList = document.querySelector('.js-btn-add');
 
-
     btnAddShopList.addEventListener('click', () => {
+        refs.containerModalShopList.innerHTML = createMarkupModalRemoveShopList(book);
+
         localStorageBooks.push(book);
         saveInLocal(API_KEY, localStorageBooks);
+
+        const btnCloseModalShopList = document.querySelector('.modal-shop-list-close')
+        btnCloseModalShopList.addEventListener('click', toggleModal);
     });
 
     // refs.btnRemoveShopList.addEventListener('click', () => {
