+++ src/partials/modal-shop-list.html
@@ -58,6 +58,7 @@
     add to shopping list
   </button>
 </div>
+</div> -->
 
 <div class="backdrop is-hidden" id="modal-remove">
   <div class="modal-shop-list" data-id="${_id}">
@@ -122,4 +123,4 @@
       press the button “Remove from the shopping list”.
     </p>
   </div>
-</div> -->
+</div>
